requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
openai==1.12.0
supabase==2.3.4
google-api-python-client==2.118.0
google-auth==2.27.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
httpx==0.26.0
python-dotenv==1.0.1
